Partially integrate uninit memory checks into `verify_std`
